requirements.txt
openai-whisper
torch
torchaudio
ffmpeg-python
pyaudio
webrtcvad
numpy
faster-whisper